Refactoring BCJFilter move to another module
also avoid multiple inheritance
